+++ R/liply.r
@@ -12,7 +12,7 @@
 #' @param ... other arguments passed on to \code{.fun}
 #' @export
 #' @examples
-#' if(require("iterators")) {
+#' if (require("iterators")) {
 #'   system.time(dlply(baseball, "id", summarise, mean_rbi = mean(rbi)))
 #'   system.time({
 #'     baseball_id <- isplit2(baseball, baseball$id)