+++ R/llply.r
@@ -72,7 +72,7 @@ llply <- function(.data, .fun = NULL, ..., .progress = "none", .inform = FALSE,
     fe_call <- as.call(c(list(as.name("foreach"), i = i), .paropts))
     fe <- eval(fe_call)
 
-    result <- fe %dopar% do.ply(i)
+    result <- foreach::`%dopar%`(fe, do.ply(i))
   } else {
     result <- loop_apply(n, do.ply)
   }