+++ R/parallel.r
@@ -1,9 +1,9 @@
 setup_parallel <- function() {
-  if (!require("foreach")) {
+  if (!requireNamespace("foreach", quietly = TRUE)) {
     stop("foreach package required for parallel plyr operation",
       call. = FALSE)
   }
-  if (getDoParWorkers() == 1) {
+  if (foreach::getDoParWorkers() == 1) {
     warning("No parallel backend registered", call. = TRUE)
   }
 }