+++ NEWS.md
@@ -1,5 +1,7 @@
 # Version 1.8.1.9000
 
+* Update for changes in R namespace best-practices.
+
 # Version 1.8.1
 
 * New parameter `.id` to `ldply()` and `rdply()` that specifies the name of